Refs 10521. Moving ChangeKind_t to sepparate header.
